Add docker-compose, celery, redis, flower, heroku.yml, e-mail
