+++ app/templates/index.html
@@ -52,6 +52,14 @@
                                 <div class="panel-footer">
                                     <input type="submit" name="submit" class="form-control" value="Scrape & Download">
                                 </div>
+                                <div class="form-group row">
+                                    <label for="recipient" class="col-sm-2 col-form-label">Email</label>
+                                    <div class="col-sm-10">
+                                        <input type="text" class="form-control" name="recipient" placeholder="email@domain.com">
+                                    </div>
+<!--                                 <div class="panel-footer">
+                                    <input type="text" name="recipient" class="form-control"> -->
+                                </div>
                                 <table class="table table-bordered table-hover" id="tableAddRow">
                                     <thead>
                                         <tr>
