+++ celery_config.py
@@ -1,0 +1,5 @@
+import os
+
+
+BROKER_URL = os.environ.get('REDIS_URL')
+CELERY_RESULT_BACKEND = os.environ.get('REDIS_URL')
