+++ config.py
@@ -10,3 +10,13 @@ basedir = os.path.abspath(os.path.dirname(__file__))
 SQLALCHEMY_DATABASE_URI = 'sqlite:///' + os.path.join(basedir, 'app.db')
 SQLALCHEMY_MIGRATE_REPO = os.path.join(basedir, 'db_repository')
 SQLALCHEMY_TRACK_MODIFICATIONS = True
+
+# Flask-Mail configuration
+MAIL_SERVER = 'smtp.gmail.com'
+MAIL_PORT = 465
+# MAIL_PORT = 587
+MAIL_USE_SSL = True
+# MAIL_USE_TLS = True
+MAIL_USERNAME = os.environ.get('MAIL_USERNAME')
+MAIL_PASSWORD = os.environ.get('MAIL_PASSWORD')
+MAIL_DEFAULT_SENDER = 'scraper@ypscraper.com'
