+++ docker-compose.yml
@@ -1,0 +1,54 @@
+version: '3.7'
+
+services:
+
+  web:
+    build: 
+      context: .
+      dockerfile: ./compose/Dockerfile
+    image: web
+    container_name: web
+    ports:
+      - 8000:8000
+    volumes:
+      - ./.:/usr/src/app
+    command: gunicorn -b 0.0.0.0:8000 app:app --worker-class gevent
+    env_file:
+      - .env
+    depends_on:
+      - redis
+    networks:
+      - dockernet
+
+  celery:
+    image: web
+    volumes:
+      - ./.:/usr/src/app
+      - ././logs:/usr/src/app/logs
+    command: celery -A app.celery worker --loglevel=info  --logfile=logs/celery.log
+    env_file:
+      - .env
+    depends_on:
+      - web
+      - redis
+    networks:
+      - dockernet
+
+  redis:
+    image: redis
+    networks:
+      - dockernet
+
+  monitor:
+    image: web
+    ports:
+      - 5555:5555
+    command:  flower -A app.celery --port=5555 --broker=redis://redis:6379/0
+    depends_on:
+      - web
+      - redis
+    networks:
+      - dockernet
+
+networks:
+    dockernet:
