+++ requirements.txt
@@ -1,37 +1,19 @@
-alembic==0.9.6
-bcrypt==3.1.4
-beautifulsoup4==4.6.0
-blinker==1.4
-certifi==2017.7.27.1
-cffi==1.11.2
-chardet==3.0.4
-click==6.7
-dominate==2.3.1
-Flask==0.12.2
-Flask-Bootstrap==3.3.7.1
-Flask-Login==0.4.0
-Flask-Mail==0.9.1
-Flask-Migrate==2.1.1
-flask-paginate==0.5.1
-Flask-SQLAlchemy==2.3.2
-Flask-User==0.6.19
-Flask-WTF==0.14.2
-gunicorn==19.7.1
-idna==2.6
-itsdangerous==0.24
-Jinja2==2.9.6
-Mako==1.0.7
-MarkupSafe==1.0
-passlib==1.7.1
-pycparser==2.18
-pycryptodome==3.4.7
-python-dateutil==2.6.1
-python-editor==1.0.3
-requests==2.18.4
-six==1.11.0
-SQLAlchemy==1.1.14
-urllib3==1.22
-visitor==0.1.3
-Werkzeug==0.12.2
-WTForms==2.1
-WTForms-SQLAlchemy==0.1
+beautifulsoup4
+celery
+Flask
+Flask-Bootstrap
+# Flask-Login  -or-  Flask-User
+Flask-Mail
+Flask-Migrate
+flask-paginate
+Flask-SQLAlchemy
+Flask-WTF
+flower
+gevent
+gunicorn
+redis
+requests
+SQLAlchemy
+Werkzeug
+WTForms
+WTForms-SQLAlchemy
